refactor(ffmpeg/transcode): split phase methods + context into sibling files
PR C of the video_convert.rs decomposition sprint. After PR B the file
was 894 LOC; this split brings the orchestrator file to ~125 LOC by
moving:
- The 3 phase-state structs (Phase1Outputs, AudioTranscodeState,
  VideoTranscodeContext) to video_transcode_context.rs (~80 LOC)
- The 6 phase methods to video_transcode_phases.rs (~709 LOC)

video_convert.rs now hosts only convert_video (async entry),
convert_video_sync (remux/transcode router), and the thin
convert_video_transcode_sync orchestrator.

Visibility lifted from fn -> pub(super) fn on the 6 phase methods.
Struct + all fields lifted to pub(super) for cross-module
construction/destructure. No helpers in video_convert.rs needed
visibility changes — all Self:: callees were already pub(super)
in sibling modules (video_pipeline.rs, audio_recode_pipeline.rs)
or pub(crate) in ffi_helpers.

Pure structural move; behavior unchanged; tests pass.
